Split index file and improve clean coding
