+++ src/core/associateActions.js
@@ -1,0 +1,12 @@
+export function associateActions(store, actions) {
+  const associatedActions = {};
+  Object.keys(actions).forEach((key) => {
+    if (typeof actions[key] === "function") {
+      associatedActions[key] = actions[key].bind(null, store);
+    }
+    if (typeof actions[key] === "object") {
+      associatedActions[key] = associateActions(store, actions[key]);
+    }
+  });
+  return associatedActions;
+}
