+++ src/core/cleanUpListener.js
@@ -1,0 +1,5 @@
+export const cleanUpListener = (store, newListener) => () => {
+  store.listeners = store.listeners.filter(
+    (listener) => listener !== newListener
+  );
+};
