+++ src/core/runListeners.js
@@ -1,0 +1,6 @@
+export const runListeners = (store, afterUpdateCallback) => {
+  store.listeners.forEach((listener) => {
+    listener.run(store.state);
+  });
+  afterUpdateCallback && afterUpdateCallback();
+};
